+++ .gitignore
@@ -146,6 +146,8 @@ venv.bak/
 
 # Visual Studio Code
 .vscode/
+!.vscode/settings.json
+!.vscode/extensions.json
 
 # Visual Studio 2015/2017/2019/2022
 .vs/
