+++ pyproject.toml
@@ -1,0 +1,53 @@
+[tool.poetry]
+name = "job-hunt-assistant"
+version = "0.1.0"
+description = "job hut assistant is a tool which is desiged to search, apply sutable jobs for you. it makes tailored application suitable for your profile and job description"
+authors = ["Akhil <surapuramakhil@gmail.com>"]
+license = "AGPL"
+readme = "README.md"
+package-mode = false
+
+[[tool.poetry.source]]
+name = "pytorch"
+url = "https://download.pytorch.org/whl/cpu"  # For CPU-only version
+priority = "explicit"
+
+[tool.poetry.dependencies]
+python = "^3.13"
+selenium = "4.9.1"
+httpx = ">=0.27.0,<0.28.0"
+inputimeout = "^1.0.4"
+jsonschema = "^4.23.0"
+jsonschema-specifications = "^2024.10.1"
+langchain = "^0.3.10"
+langchain-anthropic = "^0.3.0"
+langchain-groq = "^0.2.1"
+langchain-community = "^0.3.10"
+langchain-core = "^0.3.22"
+langsmith = ">=0.1.125,<0.2.0"
+langchain-google-genai = "^2.0.6"
+langchain-openai = "^0.2.11"
+langchain-text-splitters = "^0.3.2"
+levenshtein = "^0.26.1"
+loguru = "^0.7.3"
+openai = "^1.57.0"
+pdfminer-six = "^20240706"
+pytest = "^8.3.4"
+python-dotenv = "^1.0.1"
+pyyaml = "^6.0.2"
+regex = "^2024.7.24"
+reportlab = "^4.2.5"
+webdriver-manager = "^4.0.2"
+pytest-mock = "^3.14.0"
+pytest-cov = "^6.0.0"
+click = "^8.1.7"
+lib-resume-builder-aihawk = {git = "https://github.com/feder-cr/lib_resume_builder_AIHawk.git"}
+langchain-ollama = "^0.2.1"
+#langchain-huggingface # commented due to its dependency with torch
+
+[tool.poetry.extras]
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
