+++ src/utils/browser_utils.py
@@ -2,7 +2,7 @@
 import random
 import time
 
-from src.logging import logger
+from logger import logger
 
 
 def is_scrollable(element):
