+++ src/webdrivers/firefox.py
@@ -3,7 +3,7 @@ from selenium import webdriver
 from selenium.webdriver.firefox.service import Service as FirefoxService
 from webdriver_manager.firefox import GeckoDriverManager
 
-from src.webdrivers.base_browser import Browser
+from webdrivers.base_browser import Browser
 
 
 class Firefox(Browser):
