+++ tests/test_aihawk_bot_facade.py
@@ -1,5 +1,5 @@
 import pytest
-# from src.aihawk_job_manager import JobManager
+# from aihawk_job_manager import JobManager
 
 @pytest.fixture
 def job_manager():
