+++ tests/test_aihawk_job_manager.py
@@ -1,13 +1,13 @@
 import json
 import re
-from src.job import Job
+from job import Job
 from unittest import mock
 from pathlib import Path
 import os
 import pytest
 from ai_hawk.job_manager import AIHawkJobManager
 from selenium.common.exceptions import NoSuchElementException
-from src.logging import logger
+from logger import logger
 
 
 @pytest.fixture
